client/python/requirements.txt
pyyaml==5.1.1
requests==2.21.0
typing==3.7.4
